src/requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
attrs==18.2.0
Babel==2.6.0
beautifulsoup4==4.6.0
cached-property==1.5.1
certifi==2018.10.15
cffi==1.12.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.6.1
datapunt-authorization-django==0.2.24
datapunt-config-loader==1.1.0
datapunt-objectstore==2018.12.10
debtcollector==1.21.0
defusedxml==0.5.0
Django==2.1.7
django-cors-headers==2.4.0
django-debug-toolbar==1.10.1
django-extensions==2.1.4
django-filter==2.0.0
django-filters==0.2.1
django-modelcluster==4.3
django-rest-swagger==2.2.0
django-taggit==0.23.0
django-treebeard==4.3
djangorestframework==3.8.2
djangorestframework-csv==2.1.0
djangorestframework-gis==0.13
djangorestframework-xml==1.3.0
draftjs-exporter==2.1.5
drf-amsterdam==0.2.1
drf-extensions==0.4.0
factory-boy==2.11.1
Faker==1.0.4
flake8==3.6.0
geojson==2.4.1
html5lib==1.0.1
idna==2.7
iso8601==0.1.12
isodate==0.6.0
itypes==1.1.0
Jinja2==2.10
jsonschema==3.0.1
keystoneauth1==3.13.1
lxml==4.2.5
MarkupSafe==1.1.0
mccabe==0.6.1
msgpack==0.6.1
netaddr==0.7.19
netifaces==0.10.9
numpy==1.15.4
openapi-codec==1.3.2
os-service-types==1.6.0
oslo.config==6.8.1
oslo.i18n==3.23.1
oslo.serialization==2.28.2
oslo.utils==3.40.3
pandas==0.23.4
pbr==5.1.3
Pillow==5.4.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pycodestyle==2.4.0
pycparser==2.19
pyflakes==2.0.0
PyJWT==1.7.1
pyparsing==2.4.0
pyrsistent==0.14.11
python-dateutil==2.7.5
python-keystoneclient==3.19.0
python-swiftclient==3.7.0
pytz==2018.7
PyYAML==5.1
requests==2.20.1
requests-toolbelt==0.8.0
rfc3986==1.2.0
sentry-sdk==0.6.4
simplejson==3.16.0
six==1.11.0
sqlparse==0.2.4
stevedore==1.30.1
text-unidecode==1.2
typing==3.6.6
unicodecsv==0.14.1
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.24.1
wagtail==2.4
webencodings==0.5.1
Willow==1.1
wrapt==1.11.1
xlrd==1.1.0
zeep==3.1.0
